requirements.txt
anyio==3.6.1
arch==5.3.1
astropy==5.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cvxpy==1.2.1
cycler==0.11.0
ecos==2.0.10
fastapi==0.79.0
fonttools==4.34.4
greenlet==1.1.1
h11==0.13.0
httptools==0.4.0
idna==3.3
investpy==1.0.8
joblib==1.1.0
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.5.3
networkx==2.8.5
numpy==1.23.2
osqp==0.6.2.post5
packaging==21.3
pandas==1.4.3
patsy==0.5.2
Pillow==9.2.0
pip==22.1.2
property-cached==1.6.4
psycopg2==2.9.3
pydantic==1.9.2
pyerfa==2.0.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
qdldl==0.1.5.post2
requests==2.28.1
Riskfolio-Lib==3.3.0
scikit-learn==1.1.2
scipy==1.9.0
scs==3.2.0
setuptools==61.2.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
statsmodels==0.13.2
threadpoolctl==3.1.0
typing_extensions==4.3.0
Unidecode==1.3.4
urllib3==1.26.11
uvicorn==0.18.2
watchfiles==0.16.1
websockets==10.3
wheel==0.37.1
wincertstore==0.2
XlsxWriter==3.0.3
